refactor: introduce gallery registry
