+++ src/components/prompts/ComponentsView.tsx
@@ -4,7 +4,12 @@ import * as React from "react";
 import Fuse from "fuse.js";
 import { Card, CardContent, Skeleton, Snackbar } from "@/components/ui";
 import Badge from "@/components/ui/primitives/Badge";
-import { SPEC_DATA, type Section, type Spec } from "./constants";
+import {
+  getGalleryPreview,
+  getGallerySectionEntries,
+  type GallerySpec,
+  type Section,
+} from "./constants";
 import { cn } from "@/lib/utils";
 import { usePrefersReducedMotion } from "@/lib/useReducedMotion";
 
@@ -15,7 +20,11 @@ type ComponentsViewProps = {
   onFilteredCountChange?: (count: number) => void;
 };
 
-type SpecCardProps = Spec & {
+type ViewSpec = GallerySpec & {
+  previewNode: React.ReactNode;
+};
+
+type SpecCardProps = ViewSpec & {
   disabled?: boolean;
   onCodeVisibilityChange?: (
     specId: string,
@@ -28,7 +37,7 @@ function SpecCard({
   id,
   name,
   description,
-  element,
+  previewNode,
   props,
   code,
   disabled,
@@ -191,7 +200,7 @@ function SpecCard({
           {description}
         </p>
       ) : null}
-      <div className={frameClassName}>{element}</div>
+      <div className={frameClassName}>{previewNode}</div>
       {code ? (
         <pre
           id={codeId}
@@ -206,9 +215,9 @@ function SpecCard({
         <footer className="mt-auto">
           <ul className="flex flex-wrap items-center gap-x-[var(--space-3)] gap-y-[var(--space-2)] text-label text-muted-foreground">
             {props.map((p) => (
-              <li key={p.label} className="flex items-center gap-[var(--space-1)]">
-                <span className="font-medium text-foreground">{p.label}</span>
-                <span>{p.value}</span>
+              <li key={p.name} className="flex items-center gap-[var(--space-1)]">
+                <span className="font-medium text-foreground">{p.name}</span>
+                <span>{p.type}</span>
               </li>
             ))}
           </ul>
@@ -254,43 +263,30 @@ export default function ComponentsView({
     setActiveSpecId(null);
   }, [query, section, onCurrentCodeChange]);
 
-  const sectionSpecs = React.useMemo<Spec[]>(() => {
-    const baseSpecs = SPEC_DATA[section];
+  const sectionSpecs = React.useMemo<ViewSpec[]>(() => {
+    const rawSpecs = getGallerySectionEntries(section);
+    const baseSpecs: ViewSpec[] = rawSpecs.map((spec) => {
+      const renderPreview = getGalleryPreview(spec.preview.id);
+      const previewNode = renderPreview ? (
+        <>{renderPreview()}</>
+      ) : (
+        <div className="text-ui text-muted-foreground">Preview unavailable</div>
+      );
+
+      return {
+        ...spec,
+        previewNode,
+      };
+    });
+
     if (section !== "cards") {
       return baseSpecs;
     }
 
-    const loadingCard: Spec = {
+    const loadingCard: ViewSpec = {
       id: "card-loading-state",
       name: "Card Loading State",
       description: "Skeleton placeholders communicate asynchronous loading.",
-      element: (
-        <Card>
-          <CardContent className="space-y-[var(--space-4)]">
-            <div className="space-y-[var(--space-2)]">
-              <Skeleton
-                ariaHidden={false}
-                role="status"
-                aria-label="Loading summary"
-                className="h-[var(--space-6)] w-3/4"
-                radius="sm"
-              />
-              <Skeleton className="w-full" />
-              <Skeleton className="w-4/5" />
-            </div>
-            <div className="flex items-center gap-[var(--space-3)]">
-              <Skeleton
-                className="h-[var(--space-7)] w-[var(--space-7)] flex-none"
-                radius="full"
-              />
-              <div className="flex-1 space-y-[var(--space-2)]">
-                <Skeleton className="w-3/4" />
-                <Skeleton className="w-2/3" />
-              </div>
-            </div>
-          </CardContent>
-        </Card>
-      ),
       tags: ["card", "state", "loading"],
       code: `<Card>
   <CardContent className="space-y-[var(--space-4)]">
@@ -317,32 +313,41 @@ export default function ComponentsView({
     </div>
   </CardContent>
 </Card>`,
-    };
-
-    const errorCard: Spec = {
-      id: "card-error-state",
-      name: "Card Error State",
-      description: "Snackbar feedback surfaces failure messaging and retry.",
-      element: (
+      kind: "component",
+      preview: { id: "card-loading-state-preview" },
+      previewNode: (
         <Card>
-          <CardContent className="space-y-[var(--space-3)]">
-            <div className="space-y-[var(--space-1)]">
-              <h4 className="text-ui font-semibold tracking-[-0.01em]">
-                Data unavailable
-              </h4>
-              <p className="text-label text-muted-foreground">
-                Refresh to request the latest match insights.
-              </p>
+          <CardContent className="space-y-[var(--space-4)]">
+            <div className="space-y-[var(--space-2)]">
+              <Skeleton
+                ariaHidden={false}
+                role="status"
+                aria-label="Loading summary"
+                className="h-[var(--space-6)] w-3/4"
+                radius="sm"
+              />
+              <Skeleton className="w-full" />
+              <Skeleton className="w-4/5" />
             </div>
-            <Snackbar
-              message="Sync failed"
-              actionLabel="Retry"
-              onAction={() => {}}
-              className="mx-0 w-full justify-between border-danger/40 bg-danger/15 text-danger-foreground"
-            />
+            <div className="flex items-center gap-[var(--space-3)]">
+              <Skeleton
+                className="h-[var(--space-7)] w-[var(--space-7)] flex-none"
+                radius="full"
+              />
+              <div className="flex-1 space-y-[var(--space-2)]">
+                <Skeleton className="w-3/4" />
+                <Skeleton className="w-2/3" />
+              </div>
+            </div>
           </CardContent>
         </Card>
       ),
+    };
+
+    const errorCard: ViewSpec = {
+      id: "card-error-state",
+      name: "Card Error State",
+      description: "Snackbar feedback surfaces failure messaging and retry.",
       tags: ["card", "state", "error"],
       code: `<Card>
   <CardContent className="space-y-[var(--space-3)]">
@@ -362,6 +367,28 @@ export default function ComponentsView({
     />
   </CardContent>
 </Card>`,
+      kind: "component",
+      preview: { id: "card-error-state-preview" },
+      previewNode: (
+        <Card>
+          <CardContent className="space-y-[var(--space-3)]">
+            <div className="space-y-[var(--space-1)]">
+              <h4 className="text-ui font-semibold tracking-[-0.01em]">
+                Data unavailable
+              </h4>
+              <p className="text-label text-muted-foreground">
+                Refresh to request the latest match insights.
+              </p>
+            </div>
+            <Snackbar
+              message="Sync failed"
+              actionLabel="Retry"
+              onAction={() => {}}
+              className="mx-0 w-full justify-between border-danger/40 bg-danger/15 text-danger-foreground"
+            />
+          </CardContent>
+        </Card>
+      ),
     };
 
     return [...baseSpecs, loadingCard, errorCard];
@@ -370,7 +397,7 @@ export default function ComponentsView({
   const fuse = React.useMemo(
     () =>
       new Fuse(sectionSpecs, {
-        keys: ["name", "tags", "props.value"],
+        keys: ["name", "tags", "props.name", "props.type"],
         threshold: 0.3,
       }),
     [sectionSpecs],
