+++ types/import-meta.d.ts
@@ -1,0 +1,8 @@
+declare interface ImportMeta {
+  glob<T = unknown>(
+    pattern: string,
+    options: {
+      eager: true;
+    },
+  ): Record<string, T>;
+}
